requirements.txt

pandas>=1.5.0
scikit-learn>=1.2.0
numpy>=1.24.0
google-cloud-aiplatform>=1.35.0
google-cloud-storage>=2.10.0
kfp>=2.0.0

# Code quality tools
pylint>=3.0.0,<4.0.0
black>=23.7.0
flake8>=6.0.0

# Testing libraries
pytest>=7.4.0,<8.0.0
pytest-mock>=3.11.0,<4.0.0
pytest-cov>=4.1.0,<5.0.0

# Utilities
joblib>=1.3.0,<2.0.0